requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
cryptography==43.0.3
alembic==1.14.0
python-dateutil==2.9.0.post0
tenacity==9.0.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
